Revert "Revert "timers_master_init32(): timer period formula""
This reverts commit 5b3993ab14eda206f2aee734962ae965f10e6ace.

+++ include/timers.h
@@ -55,7 +55,7 @@ static const uint8_t TAU3offset = 0x2E;
 // Function Prototypes
 // ===========================================================
 void timers_master_init(uint32_t period_us);
-void timers_master_init32(float period_us);
+void timers_master_init32(uint32_t period_us);
 
 void timers_tau_init(	volatile uint16_t *addr_ccN,
 						volatile uint8_t *addr_ctrlb,
